git-svn-id: https://opengl-scut.googlecode.com/svn/trunk@361 81c3a017-244d-672b-1dc6-c8d6962b81c2

+++ Code/Scar/Scar/GameBag.h
@@ -40,7 +40,7 @@ namespace Network
 		float	y;
 		float	z;
 
-		AllowJoinRoomBag( int i, int xx, int yy, int zz )
+		AllowJoinRoomBag( int i, float xx, float yy, float zz )
 			: index( i ), x( xx ), y( yy ), z( zz )
 		{
 		}
